% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonAlert.R
\name{argonAlert}
\alias{argonAlert}
\title{Create a Boostrap 4 argon alert}
\usage{
argonAlert(..., icon = NULL, status = "default", closable = TRUE)
}
\arguments{
\item{...}{Alert content.}

\item{icon}{Alert icon.}

\item{status}{Alert status. See \url{https://demos.creative-tim.com/argon-design-system/docs/components/alerts.html}.}

\item{closable}{Whether to make the alert closable or not. TRUE by default.}
}
\description{
Build an argon alert
}
\examples{
if (interactive()) {
 library(argonR)
 argonAlert(
  icon = "basket",
  status = "danger",
  "This is an alert",
  closable = TRUE
 )
}


}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
