% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonBadge.R
\name{argonBadge}
\alias{argonBadge}
\title{Create a Boostrap 4 badge}
\usage{
argonBadge(text = NULL, src = NULL, pill = FALSE,
  status = "default")
}
\arguments{
\item{text}{Badge text.}

\item{src}{Badge external link.}

\item{pill}{Pill mode. FALSE by default.}

\item{status}{Badge status.}
}
\description{
Build an argon badge
}
\examples{
if (interactive()) {
}


}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
