% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-effects.R
\name{argonBlur}
\alias{argonBlur}
\title{Create a Boostrap 4 blur effect}
\usage{
argonBlur(..., text = NULL, text_color = NULL)
}
\arguments{
\item{...}{Tag to be treated by the blur effect.}

\item{text}{Text that appears on hover.}

\item{text_color}{Text color. See \url{https://demos.creative-tim.com/argon-design-system/docs/foundation/colors.html}.}
}
\description{
Build an argon blur effect
}
\note{
Not compatible with argonDash
}
\examples{
if (interactive()) {
 library(argonR)

}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
