% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonCarousel.R
\name{argonCarouselItem}
\alias{argonCarouselItem}
\title{Create a carousel item}
\usage{
argonCarouselItem(src = NULL, active = FALSE)
}
\arguments{
\item{src}{Image url or path.}

\item{active}{Whether the item is active or not.}
}
\description{
Build an argon carousel item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
