% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-effects.R
\name{argonCascade}
\alias{argonCascade}
\title{Create a Boostrap 4 cascade effect}
\usage{
argonCascade(...)
}
\arguments{
\item{...}{Slot for \link{argonCascadeItem}. Exactly 13 items.}
}
\description{
Build an argon cascade effect
}
\note{
You must pass cascade = TRUE to argonSection
}
\examples{
if (interactive()) {
 library(argonR)

}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
