% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-effects.R
\name{argonCascadeItem}
\alias{argonCascadeItem}
\title{Create a Boostrap 4 cascade item}
\usage{
argonCascadeItem(name, size = NULL, src = NULL)
}
\arguments{
\item{name}{Icon name.}

\item{size}{Icon size: NULL, "sm" or "lg".}

\item{src}{Optional external url, link, ...}
}
\description{
Build an argon cascade item
}
\examples{
if (interactive()) {
 library(argonR)

}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
