% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{argonContainer}
\alias{argonContainer}
\title{Create a boostrap 4 container}
\usage{
argonContainer(..., size = NULL)
}
\arguments{
\item{...}{Tag to be embedded.}

\item{size}{Container size. NULL or "lg".}
}
\description{
Create a boostrap 4 container
}
\examples{
if (interactive()) {
 library(argonR)

}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
