% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonNavbar.R
\name{argonDropdown}
\alias{argonDropdown}
\title{Create a Boostrap 4 argon dropdown menu}
\usage{
argonDropdown(..., name, size = NULL)
}
\arguments{
\item{...}{Slot for \link{argonDropdownItem}.}

\item{name}{Dropdown menu name.}

\item{size}{Dropdown size: NULL or "xl".}
}
\description{
Build an argon dropdown menu
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
