% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonNavbar.R
\name{argonDropdownItem}
\alias{argonDropdownItem}
\title{Create a Boostrap 4 argon dropdown item}
\usage{
argonDropdownItem(name = NULL, description = NULL, src = NULL,
  icon = NULL, status = NULL)
}
\arguments{
\item{name}{Item name.}

\item{description}{Item description if any.}

\item{src}{HTML target page.}

\item{icon}{Item icon of any.}

\item{status}{Icon and name color status. See \url{https://demos.creative-tim.com/argon-design-system/docs/foundation/colors.html}.}
}
\description{
Build an argon dropdown item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
