% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{argonH1}
\alias{argonH1}
\title{Create a custim h1 tag}
\usage{
argonH1(..., display)
}
\arguments{
\item{...}{Any text.}

\item{display}{value between 1 and 4.}
}
\description{
A custom h1 tag
}
\examples{
if (interactive()) {
 library(argonR)
 argonH1(display = 3, "ArgonR elements")
}


}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
