% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonIcon.R
\name{argonIcon}
\alias{argonIcon}
\title{Create a Nucleo icon}
\usage{
argonIcon(name, color = NULL)
}
\arguments{
\item{name}{icon name.}

\item{color}{icon color.}
}
\description{
Build an argon Nucleo icon
}
\examples{
if (interactive()) {
}


}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
