% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonImage.R
\name{argonImage}
\alias{argonImage}
\title{Create a Boostrap 4 image container}
\usage{
argonImage(src = NULL, url = NULL, floating = FALSE,
  card_mode = FALSE)
}
\arguments{
\item{src}{Image source or path.}

\item{url}{Only with card_mode on. Optional external link.}

\item{floating}{Whether to apply a floating effect. FALSE by default.}

\item{card_mode}{Whether to include the image in a card wrapper. FALSE by default.}
}
\description{
Build an argon image container
}
\examples{
if (interactive()) {
}


}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
