% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonPage.R
\name{argonInstall}
\alias{argonInstall}
\title{Install argon assets}
\usage{
argonInstall(path = getwd())
}
\arguments{
\item{path}{Where to store assets. By default, getwd().}
}
\description{
Install argon assets
}
\note{
argon assets are stored in the inst folder similarly as the package folder.
}
\examples{
if(interactive()){
 library(argonR)
 argonInstall()
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
