% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{argonLead}
\alias{argonLead}
\title{Create a custim p tag}
\usage{
argonLead(...)
}
\arguments{
\item{...}{Any text.}
}
\description{
A custom p tag
}
\examples{
if (interactive()) {
}


}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
