% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{argonMuted}
\alias{argonMuted}
\title{Create a muted text}
\usage{
argonMuted(...)
}
\arguments{
\item{...}{Any text.}
}
\description{
Apply the class muted to a text
}
\examples{
if (interactive()) {
}


}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
