% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonNavbar.R
\name{argonNavbar}
\alias{argonNavbar}
\title{Create a Boostrap 4 argon navbar}
\usage{
argonNavbar(..., src = NULL, id)
}
\arguments{
\item{...}{Slot for \link{argonNavMenu}.}

\item{src}{Brand image path or url.}

\item{id}{Navbar toggle unique id.}
}
\description{
Build an argon navbar
}
\examples{
if(interactive()){
 library(argonR)
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
