% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonNavbar.R
\name{argonNavItem}
\alias{argonNavItem}
\title{Create a Boostrap 4 argon navbar item}
\usage{
argonNavItem(name = NULL, src = NULL, icon = NULL, tooltip = NULL)
}
\arguments{
\item{name}{Item name.}

\item{src}{HTML target page.}

\item{icon}{Item icon of any.}

\item{tooltip}{Text to display when the item is hovered.}
}
\description{
Build an argon navbar item
}
\examples{
if(interactive()){
 library(argonR)
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
