% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonNavbar.R
\name{argonNavMenu}
\alias{argonNavMenu}
\title{Create a Boostrap 4 argon navbar menu}
\usage{
argonNavMenu(..., side = "left")
}
\arguments{
\item{...}{Slot for \link{argonNavItem} and \link{argonDropdown}.}

\item{side}{Nav menu side. "left by default" but can also be "right".}
}
\description{
Build an argon navbar menu
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
