% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{argonPadding}
\alias{argonPadding}
\title{Change tag padding}
\usage{
argonPadding(tag, orientation, value)
}
\arguments{
\item{tag}{Tag to be treated.}

\item{orientation}{Padding side. See \url{https://getbootstrap.com/docs/4.1/utilities/spacing/}.}

\item{value}{Padding value. See \url{https://getbootstrap.com/docs/4.1/utilities/spacing/}.}
}
\description{
Change tag padding
}
\note{
Disabled on small screens (not a bug)
}
\examples{
if (interactive()) {
 library(argonR)
 argonPadding(tag = h1("test"), orientation = NULL, value = 5)
 argonPadding(tag = h1("test"), orientation = "x", value = 2)
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
