% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonPage.R
\name{argonPage}
\alias{argonPage}
\title{Create a Boostrap 4 argon page}
\usage{
argonPage(..., title = NULL, description = NULL, author = NULL,
  navbar = NULL, footer = NULL)
}
\arguments{
\item{...}{Body content}

\item{title}{App title.}

\item{description}{Purpose.}

\item{author}{Author.}

\item{navbar}{Navbar.}

\item{footer}{Footer.}
}
\description{
Build an argon page
}
\examples{
if(interactive()){
 library(argonR)
 argonPage(
  title = "ArgonR Static Template",
  author =  "Somebody",
  description = "HTML Static Template",
  navbar = argonNavbar(id = "navbar"),
  footer = argonFooter(),
  # main content
  argonSection(),
  argonSection(),
  argonSection(),
  argonSection(),
  argonSection()
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
