% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonPage.R
\name{argonPageTemplate}
\alias{argonPageTemplate}
\title{Create an HTML version of the argonPage}
\usage{
argonPageTemplate(filename = "index.html", argonPage)
}
\arguments{
\item{filename}{HTML filename for instance, index.html.}

\item{argonPage}{Slot for \link{argonPage}.}
}
\description{
Create an HTML version of the argonPage
}
\examples{
if(interactive()){
 library(argonR)
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
