% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonPagination.R
\name{argonPagination}
\alias{argonPagination}
\title{Build an argon pagination container}
\usage{
argonPagination(..., size = NULL, align = NULL)
}
\arguments{
\item{...}{Slot for \link{argonPaginationItem}.}

\item{size}{Pagination size: NULL, "sm" or "lg".}

\item{align}{Pagination alignement. NULL (left), "center" or "end" (right).}
}
\description{
Build an argon pagination container
}
\note{
align will not work if you embed the argonPagination in an argonRow.
}
\examples{
if(interactive()){
 library(argonR)
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
