% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonPagination.R
\name{argonPaginationItem}
\alias{argonPaginationItem}
\title{Build an argon pagination item}
\usage{
argonPaginationItem(name, src = NULL)
}
\arguments{
\item{name}{Item name. Better to use numerics such as 1, 2, 3, ...}

\item{src}{Item target. Another html page for instance.}
}
\description{
Build an argon pagination item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
