% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonPersp.R
\name{argonPersp}
\alias{argonPersp}
\title{Create a Boostrap 4 perspective effect}
\usage{
argonPersp(..., side)
}
\arguments{
\item{...}{Tag to be treated by the persp effect.}

\item{side}{Perspective side effect. "right" or "left".}
}
\description{
Build an argon perspective effect
}
\note{
Disabled on small screens (not a bug)
}
\examples{
if (interactive()) {
 library(argonR)
 argonPersp(argonCard(), side = "right")
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
