% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonProfile.R
\name{argonProfile}
\alias{argonProfile}
\title{Create a Boostrap 4 profile card}
\usage{
argonProfile(..., title = NULL, subtitle = NULL, src = NULL,
  url = NULL, url_1 = NULL, url_2 = NULL, stats)
}
\arguments{
\item{...}{Any element.}

\item{title}{Profile title.}

\item{subtitle}{Profile subtitle.}

\item{src}{Profile image path or url.}

\item{url}{Main url.}

\item{url_1}{First social link.}

\item{url_2}{Second social link.}

\item{stats}{Stats items. See \link{argonProfileStats}.}
}
\description{
Build an argon profile card
}
\examples{
if (interactive()) {
}


}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
