% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonProfile.R
\name{argonProfileStat}
\alias{argonProfileStat}
\title{Create a Boostrap 4 profile card stat item}
\usage{
argonProfileStat(value, description)
}
\arguments{
\item{value}{Stat value.}

\item{description}{Stat description.}
}
\description{
Build an argon profile card stat item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
