% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonProfile.R
\name{argonProfileStats}
\alias{argonProfileStats}
\title{Create a Boostrap 4 profile card stats container}
\usage{
argonProfileStats(...)
}
\arguments{
\item{...}{Slot for \link{argonProfileStat}.}
}
\description{
Build an argon profile card stats container
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
