% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonProgress.R
\name{argonProgress}
\alias{argonProgress}
\title{Create a progress bar}
\usage{
argonProgress(value, text = "Task completed", status = "primary")
}
\arguments{
\item{value}{Progress value between 0 and 100.}

\item{text}{Label text.}

\item{status}{Progress status color. See \url{https://demos.creative-tim.com/argon-design-system/docs/foundation/colors.html}.}
}
\description{
Build an argon progress bar
}
\examples{
if (interactive()) {
 library(argonR)
 argonProgress(value = 10, status = "danger", text = "Custom Text")
}


}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
