% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{argonQuote}
\alias{argonQuote}
\title{Create a quotation tag}
\usage{
argonQuote(..., footer = NULL, source = NULL, align = "center")
}
\arguments{
\item{...}{Any text.}

\item{footer}{Footer content, if any.}

\item{source}{Source if any.}

\item{align}{Text alignement. "center", "right" or "left".}
}
\description{
A custom quotation tool
}
\examples{
if (interactive()) {
}


}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
