% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonRow.R
\name{argonRow}
\alias{argonRow}
\title{Create a Boostrap 4 row}
\usage{
argonRow(..., center = FALSE)
}
\arguments{
\item{...}{Any UI element.}

\item{center}{Whether to center row elements or not. FALSE by default.}
}
\description{
Build an argon row
}
\examples{
if (interactive()) {
}


}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
