% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonButton.R
\name{argonSocialButton}
\alias{argonSocialButton}
\title{Create a Boostrap 4 argon social button}
\usage{
argonSocialButton(src = NULL, status = "default", icon = NULL)
}
\arguments{
\item{src}{Button external link.}

\item{status}{Button color. See \url{https://demos.creative-tim.com/argon-design-system/docs/components/buttons.html}.}

\item{icon}{Button icon. Expect \link[shiny]{icon}.}
}
\description{
Build an argon social button
}
\examples{
if(interactive()){
 library(argonR)
 argonSocialButton(
  src = "http://rinterface.com",
  status = "danger",
  icon = icon("facebook")
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
