% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonTabs.R
\name{argonTab}
\alias{argonTab}
\title{Create a Boostrap 4 tab item}
\usage{
argonTab(..., tabName, active = FALSE)
}
\arguments{
\item{...}{Tab content}

\item{tabName}{Tab name: it will be also passed as the id argument. Should be unique.}

\item{active}{Whether the tab is active or not. FALSE bu default.}
}
\description{
Build an argon tab item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
