% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonTabs.R
\name{argonTabSet}
\alias{argonTabSet}
\title{Create a Boostrap 4 tabs}
\usage{
argonTabSet(..., id, card_wrapper = FALSE, horizontal = TRUE,
  circle = FALSE, size = "sm")
}
\arguments{
\item{...}{Slot for \link{argonTab}.}

\item{id}{argonTabs id. Should be unique.}

\item{card_wrapper}{Whether to embed tab content in a card. FALSE by default.}

\item{horizontal}{Whether to display tabs horizontally. TRUE by default.}

\item{circle}{Whether to display circled design. FALSE by default.}

\item{size}{Tabs size. "sm" by default. "md", "lg".}
}
\description{
Build an argon tabs
}
\examples{
if (interactive()) {
}


}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
