% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{argonTextColor}
\alias{argonTextColor}
\title{Add a custom color to a text}
\usage{
argonTextColor(tag, color)
}
\arguments{
\item{tag}{Any tag containing text.}

\item{color}{Text color.}
}
\description{
A custom color fo your text
}
\examples{
if (interactive()) {
 library(argonR)
 library(htmltools)
 argonTextColor(
  tag = h1("test"),
  color = "success"
 )
}


}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
