% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{argonTooltip}
\alias{argonTooltip}
\title{Create a boostrap 4 tooltip}
\usage{
argonTooltip(tag, position, title = NULL)
}
\arguments{
\item{tag}{Tag to be documented.}

\item{position}{Where to display the toolptip: "top", "right", "bottom" or "left".}

\item{title}{Tooltip title.}
}
\description{
Create a boostrap 4 tooltip
}
\examples{
if (interactive()) {
 library(argonR)

}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
