% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonUser.R
\name{argonUser}
\alias{argonUser}
\title{Create a Boostrap 4 user image}
\usage{
argonUser(..., title = NULL, subtitle = NULL, src = NULL)
}
\arguments{
\item{...}{Any element such as \link{argonIcon}...}

\item{title}{User title.}

\item{subtitle}{User subtitle.}

\item{src}{User image url or path.}
}
\description{
Build an argon user image
}
\examples{
if (interactive()) {
}


}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
