% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{NS}
\alias{NS}
\alias{ns.sep}
\title{Namespaced IDs for inputs/outputs}
\format{
An object of class \code{character} of length 1.
}
\usage{
NS(namespace, id = NULL)

ns.sep
}
\arguments{
\item{namespace}{The character vector to use for the namespace. 
This can have any length, though a single element is most common. 
Length 0 will cause the id to be returned without a namespace, and 
length 2 will be interpreted as multiple namespaces, in increasing 
order of specificity (i.e. starting with the top-level namespace).}

\item{id}{The id string to be namespaced (optional).}
}
\description{
Namespaced IDs for inputs/outputs
}
\keyword{datasets}
\keyword{internal}
