% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonCarousel.R
\name{argonCarousel}
\alias{argonCarousel}
\title{Create a carousel}
\usage{
argonCarousel(..., id, width = 6)
}
\arguments{
\item{...}{Slot for \link{argonCarouselItem}.}

\item{id}{Carousel unique id.}

\item{width}{Carousel width.}
}
\description{
Build an argon carousel
}
\examples{
if (interactive()) {
}


}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
