% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonColumn.R
\name{argonColumn}
\alias{argonColumn}
\title{Create a Boostrap 4 column}
\usage{
argonColumn(..., width = NULL, center = FALSE)
}
\arguments{
\item{...}{Any UI element.}

\item{width}{Optional. Column width between 1 and 12.}

\item{center}{Whether to center column elements or not. FALSE by default.}
}
\description{
Build an argon column
}
\examples{
if (interactive()) {
 library(argonR)
 argonColumn(
  width = 4,
  argonAlert(
    icon = "basket",
    status = "danger",
    "This is an alert",
    closable = TRUE
  )
 )
}


}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
