% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonColumn.R
\name{argonColumn}
\alias{argonColumn}
\title{Create a Boostrap 4 column}
\usage{
argonColumn(width = NULL)
}
\arguments{
\item{width}{Optional. Column width between 1 and 12.}
}
\description{
Build an argon column
}
\examples{
if (interactive()) {
}


}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
