% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonFooter.R
\name{argonFooter}
\alias{argonFooter}
\title{Create a Boostrap 4 argon footer}
\usage{
argonFooter(..., has_card = FALSE)
}
\arguments{
\item{...}{Footer content.}

\item{has_card}{Enable card rendering in the footer. FALSE by default.}
}
\description{
Build an argon footer
}
\examples{
if(interactive()){
 library(argonR)
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
