% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonProfile.R
\name{argonProfile}
\alias{argonProfile}
\title{Create a Boostrap 4 profile card}
\usage{
argonProfile(..., title = NULL, subtitle = NULL, src = NULL,
  url = NULL, url_1 = NULL, url_2 = NULL, stats)
}
\arguments{
\item{...}{Any element.}

\item{title}{Profile title.}

\item{subtitle}{Profile subtitle.}

\item{src}{Profile image path or url.}

\item{url}{Main url.}

\item{url_1}{First social link.}

\item{url_2}{Second social link.}

\item{stats}{Stats items. See \link{argonProfileStats}.}
}
\description{
Build an argon profile card
}
\note{
Do not work on argonR. Only for argon Dash
}
\examples{
if(interactive()){
 library(argonR)
 argonProfile(
  title = "John",
  subtitle = "Japan, Kagoshima",
  src = "https://image.flaticon.com/icons/svg/1006/1006540.svg",
  url = "https://www.google.com",
  url_1 = "https://www.google.com",
  url_2 = "https://www.google.com",
  stats = argonProfileStats(
    argonProfileStat(
      value = 22,
      description = "Friends"
    ),
    argonProfileStat(
      value = 10,
      description = "Photos"
    ),
    argonProfileStat(
      value = 89,
      description = "Comments"
    )
  ),
  "An artist of considerable range, Ryan — 
  the name taken by Melbourne-raised, 
  Brooklyn-based Nick Murphy — writes, 
  performs and records all of his own music, 
  giving it a warm, intimate feel with a solid 
  groove structure. An artist of considerable 
  range."
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
