% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonSection.R
\name{argonSection}
\alias{argonSection}
\title{Create a Boostrap 4 section container}
\usage{
argonSection(..., size = NULL, status = "default", gradient = FALSE,
  separator = FALSE)
}
\arguments{
\item{...}{Any UI element.}

\item{size}{Section size. NULL by default.}

\item{status}{Section status.}

\item{gradient}{Section gradient effect. FALSE by default.}

\item{separator}{Section bottom separator. FALSE by default.}
}
\description{
Build an argon section container
}
\examples{
if (interactive()) {
}


}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
