% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{argonTooltip}
\alias{argonTooltip}
\title{Create a boostrap 4 tooltip}
\usage{
argonTooltip(tag, position, title = NULL)
}
\arguments{
\item{tag}{Tag to be documented.}

\item{position}{Where to display the toolptip: "top", "right", "bottom" or "left".}

\item{title}{Tooltip title.}
}
\description{
Create a boostrap 4 tooltip
}
\examples{
if (interactive()) {
 library(argonR)
 argonTooltip(
  position = "left", 
  title = "I am a nice card",
  argonCard(
   status = "primary",
   width = 12,
   title = "Card 1",
   hover_lift = TRUE,
   shadow = TRUE,
   icon = "check-bold",
   src = "#",
   "Argon is a great free UI package based on Bootstrap 4 
   that includes the most important components and features."
  )
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
